P1-2: Add absolute byte totals to AllocationPatternDomainResult
- Add TotalManagedBytes, Gen0Bytes, Gen1Bytes, Gen2Bytes, LohBytes fields to result
- Compute absolute byte totals in analyzer alongside percentages
- Surface totals in AllocationPatternSectionBuilder key metrics with FormatBytes helper
- Enables severity assessment without relying on percentages alone
- Also fixes P0-1 fallout: update reporting section builders for long gen counts
  - TypeSystemSectionBuilder: FinalizableOverheadCandidate, GenRatio helpers
  - GCPressureSectionBuilder: gen count accumulation

Addresses audit gap: "No absolute byte reporting. An engineer cannot tell whether
3% Gen2 represents 30 MB or 30 GB."

+++ src/DumpDetective.Reporting/SectionBuilders/GCPressureSectionBuilder.cs
@@ -94,7 +94,7 @@ internal sealed class GCPressureSectionBuilder : SectionBuilderBase, IAnalyzerSe
             for (int i = 0; i < limit; i++)
             {
                 TypeGenerationProfile p = d.PerTypeGenerationProfiles[i];
-                int total = p.Gen0Count + p.Gen1Count + p.Gen2Count + p.LohCount;
+                long total = p.Gen0Count + p.Gen1Count + p.Gen2Count + p.LohCount;
                 double gen2Pct   = total == 0 ? 0.0 : p.Gen2Count  * 100.0 / total;
                 double survivalR = total == 0 ? 0.0 : (p.Gen2Count + p.LohCount) * 1.0 / total;
                 rows.Add(Row(
